Add forex data request to app
To allow the API to convert prices into a foreign currency external
rates are required. This should be saved locally and used to lookup
the exchange rate.

As the this is an external API call a promise is required. Mocking a
promise with Jest and Supertest is currently not working as intended and
so an integration test will need to be skipped.