Fix Sale pricing to return in GBP or specified currency
Pricing currently occurs in pennies, before being returned or converted 
to another currency, which will then also be in the penny equivalent. 
This is not clear in the returned data, as it just declares a currency 
at the GBP equivalent level.

Fix returned pricing so that prices, totals and vat return at the 
correct stated level of the currency code, to two decimal places.